Fix mctpd PID masks to include RSVD field (#194)
CPU and Non legacy BMC were having same instance id and mctpd was not
able to differentiated between them. As a workaround CPU will have 0xF
and NLBMC's will have 0xB in reserved portion of PID.

Add reserved field also in PID mask to identify NLBMC and CPU

+++ include/hw/aspeed/I3CDriver.hpp
@@ -34,7 +34,7 @@ class I3CDriver : public hw::I3CDriver
 {
   public:
     I3CDriver(boost::asio::io_context& ioc, uint8_t i3cBusNum,
-              std::optional<uint8_t> cpuPidMask = 0);
+              std::optional<uint16_t> cpuPidMask = 0);
     ~I3CDriver() override;
 
     void init() override;
@@ -50,7 +50,7 @@ class I3CDriver : public hw::I3CDriver
     mctp_binding_asti3c* i3c{};
     bool isController = false;
     std::string i3cDeviceFile;
-    std::optional<uint8_t> pidMask;
+    std::optional<uint16_t> pidMask;
     uint8_t busNum;
     void discoverI3CDevices();
     void rescanI3CBus();
