+++ include/utils/Configuration.hpp
@@ -75,7 +75,7 @@ struct I3CConfiguration : Configuration
     uint8_t bus;
     uint8_t I3CAddress = 0;
     bool requiresCpuPidMask = true;
-    uint8_t provisionalIdMask;
+    uint16_t provisionalIdMask;
     uint16_t getRoutingInterval = 0;
     bool forwaredEIDPoolToEP = false;
     bool blockDiscoveryNotify = false;
