+++ include/utils/i3c_utils.hpp
@@ -21,7 +21,7 @@
 
 std::vector<std::string> getDevFilePaths(std::string& matchString);
 
-bool findMCTPI3CDevice(uint8_t busNum, std::optional<uint8_t> pidMask,
+bool findMCTPI3CDevice(uint8_t busNum, std::optional<uint16_t> pidMask,
                        std::string& file);
 
 bool getPID(std::string& path, std::string& pidStr);
